PyroPanel/requirements.txt
# Web Framework
fastapi>=0.103.0
uvicorn>=0.23.2
jinja2>=3.1.2
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.20
alembic>=1.12.0
psycopg2-binary>=2.9.7  # For PostgreSQL
pymysql>=1.1.0  # For MySQL

# Authentication
python-jose>=3.3.0  # For JWT
passlib>=1.7.4
bcrypt>=4.0.1

# Docker Management
docker>=6.1.3

# File Management
paramiko>=3.3.1  # For SFTP

# Task Scheduling
celery>=5.3.4
redis>=5.0.0

# Utilities
pydantic>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
